Particle Deposition Added
The user can now run the particle deposition algorithm the same way as the faulting algo. The can add single particles, run the algo indefinitely or choose to run it for a set amount of iterations. The particle deposition also works along side the faulting, with both being able to be run concurrently.