Height Based Texturing Added
The terrain now textures with 3 textures, namely, sand, grass, and snow. These are textured in their respective order according to the height of the terrain. This required some research on how to effectively map the boundary range between the textures to the 0 - 1 range for the hlsl lerp function coefficient, otherwise unsightly banding was produced as it was textured from one to the next. The effect is not perfect but there is now smoothing between textures as the height increases.